chore: move caption icon to its own module
