+++ apps-rendering/src/components/MainMedia/ImmersiveCaption.tsx
@@ -2,10 +2,9 @@
 
 import type { SerializedStyles } from '@emotion/react';
 import { css } from '@emotion/react';
-import {
-	CaptionIcon,
-	IconVariant,
-} from '@guardian/common-rendering/src/components/figCaption';
+import CaptionIcon, {
+	CaptionIconVariant,
+} from '@guardian/common-rendering/src/components/captionIcon';
 import { text } from '@guardian/common-rendering/src/editorialPalette';
 import type { ArticleFormat } from '@guardian/libs';
 import { from, remSpace, textSans } from '@guardian/source-foundations';
@@ -61,7 +60,7 @@ const ImmersiveCaption: FC<Props> = ({ mainMedia, format }) =>
 		return (
 			<p id={immersiveCaptionId} css={styles(format)}>
 				<CaptionIcon
-					variant={IconVariant.Image}
+					variant={CaptionIconVariant.Image}
 					format={format}
 					supportsDarkMode={true}
 				/>
