+++ common-rendering/src/components/captionIcon.stories.tsx
@@ -5,13 +5,12 @@
 import { ArticleDesign, ArticleDisplay, ArticlePillar } from '@guardian/libs';
 import { some } from '@guardian/types';
 import type { FC } from 'react';
-import { CaptionIconVariant } from './captionIcon';
-import FigCaption from './figCaption';
+import CaptionIcon, { CaptionIconVariant } from './captionIcon';
 
 // ----- Stories ----- //
 
 const Image: FC = () => (
-	<FigCaption
+	<CaptionIcon
 		format={{
 			design: ArticleDesign.Standard,
 			display: ArticleDisplay.Standard,
@@ -23,11 +22,11 @@ const Image: FC = () => (
 		{some(
 			'Age of the train … a tourist train in Switzerland. Photograph: Kisa_Markiza/Getty Images',
 		)}
-	</FigCaption>
+	</CaptionIcon>
 );
 
 const Video: FC = () => (
-	<FigCaption
+	<CaptionIcon
 		format={{
 			design: ArticleDesign.Standard,
 			display: ArticleDisplay.Standard,
@@ -39,14 +38,14 @@ const Video: FC = () => (
 		{some(
 			'Age of the train … a tourist train in Switzerland. Photograph: Kisa_Markiza/Getty Images',
 		)}
-	</FigCaption>
+	</CaptionIcon>
 );
 
 // ----- Exports ----- //
 
 export default {
-	component: FigCaption,
-	title: 'Common/Components/FigCaption',
+	component: CaptionIcon,
+	title: 'Common/Components/CaptionIcon',
 };
 
 export { Image, Video };
