+++ common-rendering/src/components/figCaption.tsx
@@ -5,71 +5,14 @@ import { css } from '@emotion/react';
 import { remSpace } from '@guardian/source-foundations';
 import { neutral } from '@guardian/source-foundations';
 import { textSans } from '@guardian/source-foundations';
-import { SvgCamera, SvgVideo } from '@guardian/source-react-components';
 import type { ArticleFormat } from '@guardian/libs';
 import { ArticleDesign } from '@guardian/libs';
 import type { Option } from '@guardian/types';
 import { OptionKind } from '@guardian/types';
 import type { FC, ReactNode } from 'react';
 import { darkModeCss } from '../lib';
-import { fill, text } from '@guardian/common-rendering/src/editorialPalette';
-
-// ----- Sub-Components ----- //
-
-enum IconVariant {
-	Image,
-	Video,
-}
-
-interface IconProps {
-	format: ArticleFormat;
-	supportsDarkMode: boolean;
-	variant: IconVariant;
-}
-
-const iconStyles = (supportsDarkMode: boolean): SerializedStyles => css`
-	display: inline-block;
-	width: 1.2rem;
-	margin-right: ${remSpace[1]};
-	fill: ${fill.cameraCaptionIcon()};
-	position: relative;
-
-	::before {
-		content: ' ';
-		display: block;
-		padding-top: 0.85rem;
-	}
-
-	svg {
-		position: absolute;
-		top: 0;
-		left: 0;
-		width: 100%;
-	}
-
-	${darkModeCss(supportsDarkMode)`
-        fill: ${fill.cameraCaptionIconDark()};
-    `}
-`;
-
-const Icon: FC<IconProps> = ({ format, supportsDarkMode, variant }) => {
-	switch (format.design) {
-		case ArticleDesign.Gallery:
-		case ArticleDesign.Audio:
-		case ArticleDesign.Video:
-			return null;
-		default:
-			return (
-				<span css={iconStyles(supportsDarkMode)}>
-					{variant === IconVariant.Image ? (
-						<SvgCamera />
-					) : (
-						<SvgVideo />
-					)}
-				</span>
-			);
-	}
-};
+import { text } from '@guardian/common-rendering/src/editorialPalette';
+import CaptionIcon, { CaptionIconVariant } from './captionIcon';
 
 // ----- Component ----- //
 
@@ -79,7 +22,7 @@ type Props = {
 	children: Option<ReactNode>;
 	className?: string;
 	css?: SerializedStyles;
-	variant?: IconVariant;
+	variant?: CaptionIconVariant;
 };
 
 const styles = (format: ArticleFormat, supportsDarkMode: boolean) => css`
@@ -122,7 +65,7 @@ const FigCaption: FC<Props> = ({
 	supportsDarkMode,
 	children,
 	className,
-	variant = IconVariant.Image,
+	variant = CaptionIconVariant.Image,
 }) => {
 	switch (children.kind) {
 		case OptionKind.Some:
@@ -131,7 +74,7 @@ const FigCaption: FC<Props> = ({
 					className={className}
 					css={getStyles(format, supportsDarkMode)}
 				>
-					<Icon
+					<CaptionIcon
 						format={format}
 						supportsDarkMode={supportsDarkMode}
 						variant={variant}
@@ -148,5 +91,3 @@ const FigCaption: FC<Props> = ({
 // ----- Exports ----- //
 
 export default FigCaption;
-
-export { IconVariant, Icon as CaptionIcon };
